Refactor opset version types from i64 to usize
Changed the opset version type from i64 to usize across proto_conversion, pipeline, util, and post_processing modules for consistency and type safety. Updated related function signatures, enum variants, and conversion logic to use usize, ensuring proper handling of ONNX opset versions.

+++ crates/onnx-ir/src/phases/post_processing.rs
@@ -193,7 +193,7 @@ pub(crate) fn post_process(
             // Constant lifting is a best-effort optimization after identity elimination.
             // Not all arguments can be lifted (e.g., already Static, Dynamic), so we log
             // errors but don't fail the pipeline.
-            if let Err(e) = processor.lift_constants(node, MIN_OPSET_VERSION as usize) {
+            if let Err(e) = processor.lift_constants(node, MIN_OPSET_VERSION) {
                 log::debug!(
                     "Could not lift constants for node '{}' (type: {:?}): {:?}",
                     node.name,
